requirements.txt
matlabengine~=24.2
pandas~=2.2.3
matplotlib~=3.9.2
seaborn~=0.13.2
requests~=2.32.3
gdown~=5.2.0
tqdm~=4.66.5
torchvision~=0.18.1
pillow~=11.0.0
numpy~=1.26.4
pytorch~=2.3.1
pyiqa~=0.1.11
scipy~=1.14.1